move mnist.py to datasets subfolder
