+++ .gitignore
@@ -129,5 +129,4 @@ ENV/
 
 # data
 .data/
-datasets/
 mnist/
