+++ tests/datasets/mnist.py
@@ -62,7 +62,7 @@ class MNIST(Dataset):
 
     @property
     def processed_folder(self):
-        return os.path.join(self.root, self.__class__.__name__, 'processed')
+        return os.path.join(self.root, 'mnist', 'processed')
 
     def _check_exists(self):
         train_file = os.path.join(self.processed_folder, self.training_file)
